int atomicio -> ssize_t (for alpha). ok deraadt@

+++ src/usr.bin/ssh/ssh.h
@@ -687,7 +687,7 @@ struct envstring {
 /*
  * Ensure all of data on socket comes through. f==read || f==write
  */
-int	atomicio(int (*f)(), int fd, void *s, size_t n);
+ssize_t	atomicio(ssize_t (*f)(), int fd, void *s, size_t n);
 
 #ifdef KRB4
 #include <krb.h>
